+++ src/client/java/com/zenil/coordsonscreen/CoordsOnScreen.java
@@ -1,7 +1,5 @@
 package com.zenil.coordsonscreen;
 
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendering.v1.hud.HudElementRegistry;
 import net.minecraft.resources.Identifier;
@@ -12,7 +10,7 @@ public class CoordsOnScreen implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
-        AutoConfig.register(CoordsConfig.class, GsonConfigSerializer::new);
+        CoordsConfig.load();
         HudElementRegistry.addLast(
             Identifier.fromNamespaceAndPath(MOD_ID, "coords"),
             new CoordsHudElement()